Separate attachment delete guard protocol
Move delete preparation, commit recovery, encrypted guard files, and orphan reconciliation into a focused runtime module while preserving the existing command API.
